Don't monitor read only file system

+++ agent/sysinfo/fsstat.go
@@ -56,7 +56,7 @@ func FsRWMetrics() (L []*common.Metric) {
 		file := filepath.Join(du.FsFile, ".loda-fs-detect")
 		now := time.Now().Format("2006-01-02 15:04:05")
 		content := "FS-RW" + now
-		err = CheckFS(file, content)
+		err = CheckFS(file, content, mountPoints[idx][3])
 		if err != nil {
 			res = 0
 		} else {
@@ -69,35 +69,39 @@ func FsRWMetrics() (L []*common.Metric) {
 	return
 }
 
-func CheckFS(file string, content string) error {
-	//write test
-	fd, err := os.OpenFile(file, os.O_RDWR|os.O_CREATE, 0644)
-	defer fd.Close()
-	if err != nil {
-		log.Error("Open file failed: ", err)
-		return err
-	}
-	buf := []byte(content)
-	count, err := fd.Write(buf)
-	if err != nil || count != len(buf) {
-		log.Error("Write file failed: ", err)
-		return err
-	}
-	//read test
-	read, err := ioutil.ReadFile(file)
-	if err != nil {
-		log.Error("Read file failed: ", err)
-		return err
-	}
-	if string(read) != content {
-		log.Error("Read content failed: ", string(read))
-		return errors.New("Read content failed")
-	}
-	//clean the file
-	err = os.Remove(file)
-	if err != nil {
-		log.Error("Remove file filed: ", err)
-		return err
+func CheckFS(file string, content string, t string) error {
+	//  var t from /proc/mounts
+	//  We can not check read only file system, we can not write a file
+	if t == "rw" {
+		//write test
+		fd, err := os.OpenFile(file, os.O_RDWR|os.O_CREATE, 0644)
+		defer fd.Close()
+		if err != nil {
+			log.Error("Open file failed: ", err)
+			return err
+		}
+		buf := []byte(content)
+		count, err := fd.Write(buf)
+		if err != nil || count != len(buf) {
+			log.Error("Write file failed: ", err)
+			return err
+		}
+		//read test
+		read, err := ioutil.ReadFile(file)
+		if err != nil {
+			log.Error("Read file failed: ", err)
+			return err
+		}
+		if string(read) != content {
+			log.Error("Read content failed: ", string(read))
+			return errors.New("Read content failed")
+		}
+		//clean the file
+		err = os.Remove(file)
+		if err != nil {
+			log.Error("Remove file filed: ", err)
+			return err
+		}
 	}
 	return nil
 }
